添加用户自服务管理，用户可以自己修改用户昵称和密码，添加这个模块作用是给用户一个 自己修改自己基本信息的场所，防止用户无法修改自身基本信息。

+++ crawler-backend/src/main/java/com/crawler/service/impl/UserServiceImpl.java
@@ -95,13 +95,16 @@ public class UserServiceImpl implements UserService {
         sysUser.setPassword(password);
         // 首先更新用户表
         userMapper.userUpdate(sysUser);
-        // 然后将用户角色表中数据根据用户id清空
-        int userId = sysUser.getId();
-        userMapper.deleteUserRoleMapping(userId);
-        // 最后如果用户选择了角色，像用户角色表中增加数据
-        // 得到用户选择的角色
-        String userRoleStr = sysUser.getUserRoleStr();
-        addSysUserRole(userRoleStr, sysUser);
+        // 如果不是简单需改用户信息
+        if(sysUser.getSimpleUpdate() == 0) {
+            // 然后将用户角色表中数据根据用户id清空
+            int userId = sysUser.getId();
+            userMapper.deleteUserRoleMapping(userId);
+            // 最后如果用户选择了角色，像用户角色表中增加数据
+            // 得到用户选择的角色
+            String userRoleStr = sysUser.getUserRoleStr();
+            addSysUserRole(userRoleStr, sysUser);
+        }
     }
 
     /**
